refactor: split backup types
